RelativisticDynamics.jl documentation

Welcome to the documentation for RelativisticDynamics.jl, a relativistic orbital dynamics model for simulating spinning objects in curved spacetime.

Overview

RelativisticDynamics.jl is a numerical model for determining the spin-orbital dynamics of a spinning body, such as a pulsar, on a background Kerr spacetime. The code solves a set of ODEs numerically. These equations are based on the original works of Mathisson 1937, Papapetrou 1951 and Dixon 1964. Consequently these equations are known as the MPD equations. More recent works can be found in Mashoon & Singh 2006, Singh 2005, Singh, Wu & Sarty 2014 and Li,Wu & Singh 2019. Additional interesting discussion on the motion of extended bodies in GR can be found in Costa & Natário, 2015

Manual Outline

Please see the following pages of the documentation for more details

How to run RelativisticDynamics.jl

MPD Equations

Index of functions

Scope

Installation

Acknowledgements

This work solving the MPD equations was originally motivated through interesting discussions with Kinwah Wu. The port to a modern, precision-flexible model in Julia was heavily inspired by Milan Klöwer. Huge thanks to both.

Contributions are always welcome - just open a pull request

THIS HEADING

THAT HEADING

SOMETHING ELSE

RelativisticDynamics.Constants

RelativisticDynamics.Constants

RelativisticDynamics.Model

RelativisticDynamics.PrognosticVariables

RelativisticDynamics.SystemParameters

RelativisticDynamics.ELQ

RelativisticDynamics.Kretschmann_scalar

RelativisticDynamics.PlotTrajectory

RelativisticDynamics.bounds_checks

RelativisticDynamics.christoffel

RelativisticDynamics.contravariant_metric

RelativisticDynamics.convert_to_covariant

RelativisticDynamics.covariant_metric

RelativisticDynamics.covariant_minkowski

RelativisticDynamics.delta

RelativisticDynamics.initial_conditions

RelativisticDynamics.mapping_d

RelativisticDynamics.mapping_f

RelativisticDynamics.mapping_g

RelativisticDynamics.mapping_h

RelativisticDynamics.orbit

RelativisticDynamics.orbit

RelativisticDynamics.riemann

RelativisticDynamics.schwarzchild_covariant_riemann

RelativisticDynamics.sigma

RelativisticDynamics.spintensor

RelativisticDynamics.timestepping

RelativisticDynamics.Constants — Type

C = Constants(P) A struct to hold all variables which are constant over the course of the integration. These are derived from the user-defined parameters

RelativisticDynamics.Constants — Method

Generator function for a Constants struct.

RelativisticDynamics.Model — Type

RelativisticDynamics.PrognosticVariables — Type

Struct holding the so-called 'prognostic' variables

RelativisticDynamics.SystemParameters — Type

P = Parameters(kwargs...) A struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.

RelativisticDynamics.ELQ — Method

Calculate the energy, angular momentum and Carter constant

RelativisticDynamics.Kretschmann_scalar — Method

Kretschman scalar for the Kerr metric

RelativisticDynamics.PlotTrajectory — Function

Plot trajectory of a body. Assumes coordinates are Boyer Lindquist. Plots in either 2D or 3D depending on specification of dimensions. Saves a low resolution PNG figure to disk in example_media/

RelativisticDynamics.bounds_checks — Method

Look before you leap

RelativisticDynamics.christoffel — Method

map = gridded(  alms::AbstractMatrix;
                recompute_legendre::Bool=true,
                grid::Type{<:AbstractGrid}=FullGaussianGrid)

Spectral transform (spectral to grid space) from spherical coefficients alms to a newly allocated gridded field map. Based on the size of alms the grid type grid, the spatial resolution is retrieved based on the truncation defined for grid. SpectralTransform struct S is allocated to execute gridded(alms,S).

RelativisticDynamics.contravariant_metric — Method

Construct the NxN matrix of the contravariant metric. Metric components are defined via indvidual functions to allow for auto diff in unit tests

RelativisticDynamics.convert_to_covariant — Method

Convert a vector from contravariant form to convariant form using the covariant metric

RelativisticDynamics.covariant_metric — Method

Construct the NxN matrix of the covariant metric. Metric components are defined via indvidual functions to allow for auto diff in unit tests

RelativisticDynamics.covariant_minkowski — Method

m = roundup_fft(n::Int;
                small_primes::Vector{Int}=[2,3,5])

Returns an integer m >= n with only small prime factors 2, 3, 5 (default, others can be specified with the keyword argument small_primes) to obtain an efficiently fourier-transformable number of longitudes, m = 2^i * 3^j * 5^k >= n, with i,j,k >=0.

RelativisticDynamics.delta — Method

Δ = delta(r,a) The well-known delta function of the Kerr metric

RelativisticDynamics.initial_conditions — Method

Setup the initial conditions for the MPD orbital dynamics

RelativisticDynamics.mapping_d — Method

d = mapping_d(r,a,zminus) Mapping function d used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_f — Method

f = mapping_f(r,a,zminus) Mapping function f used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_g — Method

g = mapping_g(r,a) Mapping function g used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_h — Method

h = mapping_h(r,a,zminus) Mapping function h used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.orbit — Method

prognvars = runspeedy(NF,kwargs...) Runs SpeedyWeather.jl with number format NF and any additional parameters in the keyword arguments kwargs.... Any unspecified parameters will use the default values as defined in src/parameters.jl.

RelativisticDynamics.riemann — Method

Riemann tensor components of the Kerr metric. First index is the contravariant, others are covariant

RelativisticDynamics.schwarzchild_covariant_riemann — Method

Special case - the fully covariant components of the Riemann tensor for schwarzchild. From https://arxiv.org/pdf/0904.4184.pdf

RelativisticDynamics.sigma — Method

Σ = sigma(r,θ,a) The well-known sigma function of the Kerr metric

RelativisticDynamics.spintensor — Method

Calculate the contravariant spin tensor. See e.g. https://mathworld.wolfram.com/PermutationTensor.html

RelativisticDynamics.timestepping — Method

The timesteppig Integration