Index of all functions

RelativisticDynamics.Constants

RelativisticDynamics.Constants

RelativisticDynamics.Model

RelativisticDynamics.PrognosticVariables

RelativisticDynamics.SystemParameters

RelativisticDynamics.ELQ

RelativisticDynamics.Kretschmann_scalar

RelativisticDynamics.PlotTrajectory

RelativisticDynamics.StackedPlot

RelativisticDynamics.bounds_checks

RelativisticDynamics.calculate_levi

RelativisticDynamics.christoffel

RelativisticDynamics.contravariant_metric

RelativisticDynamics.convert_to_covariant

RelativisticDynamics.covariant_metric

RelativisticDynamics.delta

RelativisticDynamics.initial_conditions

RelativisticDynamics.mapping_d

RelativisticDynamics.mapping_f

RelativisticDynamics.mapping_g

RelativisticDynamics.mapping_h

RelativisticDynamics.metric_contra_g11

RelativisticDynamics.metric_contra_g14

RelativisticDynamics.metric_contra_g22

RelativisticDynamics.metric_contra_g33

RelativisticDynamics.metric_contra_g44

RelativisticDynamics.metric_g11

RelativisticDynamics.metric_g14

RelativisticDynamics.metric_g22

RelativisticDynamics.metric_g33

RelativisticDynamics.metric_g44

RelativisticDynamics.orbit

RelativisticDynamics.permutation_tensor

RelativisticDynamics.riemann

RelativisticDynamics.schwarzchild_covariant_riemann

RelativisticDynamics.sigma

RelativisticDynamics.spintensor

RelativisticDynamics.timestepping

RelativisticDynamics.Constants — Type

C = Constants(P)

A struct to hold all variables which are constant over the course of the integration. These are derived from the user-defined parameters

RelativisticDynamics.Constants — Method

Generator function for a Constants struct.

RelativisticDynamics.Model — Type

M = Model(P,C)

The model struct which holds all the parameters (P) and constants (C)

RelativisticDynamics.PrognosticVariables — Type

Struct holding the so-called 'prognostic' variables. 'Prognostic' terminology is borrowed from cliamte science where it refers to any variables that are predicted via integration

RelativisticDynamics.SystemParameters — Type

P = Parameters(kwargs...)

A struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.

RelativisticDynamics.ELQ — Method

E,L,Q = ELQ(a,α,e,ι,D)

Calculate the energy, angular momentum and Carter constant given the Keplerian orbital parameters, and the BH spin/direction

RelativisticDynamics.Kretschmann_scalar — Method

K =  Kretschmann_scalar(r,θ,a)

Kretschman scalar for the Kerr metric

RelativisticDynamics.PlotTrajectory — Function

PlotTrajectory(solution,model,dimensions=[1,2,3],savepath="")

Plot trajectory of a body. Assumes coordinates are Boyer Lindquist. Plots in either 2D or 3D depending on specification of dimensions. Saves a low resolution PNG figure to disk in example_media/

RelativisticDynamics.StackedPlot — Function

StackedPlot(solution,model,savepath="")

Plot the x-y and x-z trajectory of a body on two separate subplots

RelativisticDynamics.bounds_checks — Method

bounds_checks(P)

Look before you leap - are the user-specified kwargs in parameters physical and reasonable?

RelativisticDynamics.calculate_levi — Method

l = calculate_levi(NF)

Determine the Levi-civita psuedo tensor

RelativisticDynamics.christoffel — Method

Γ = christoffel(coords,a)

The christoffel symbols of the Kerr metric. See e.g. https://arxiv.org/pdf/0904.4184.pdf

RelativisticDynamics.contravariant_metric — Method

g=contravariant_metric(coords,a)

Construct the NxN matrix of the contravariant metric. Metric components are defined via indvidual functions to allow for auto diff in unit tests

RelativisticDynamics.convert_to_covariant — Method

p_{μ} = convert_to_covariant(metric,p^{μ})

Convert a vector from contravariant form to convariant form using the covariant metric

RelativisticDynamics.covariant_metric — Method

g=covariant_metric(coords,a)

Construct the NxN matrix of the covariant metric.

RelativisticDynamics.delta — Method

Δ = delta(r,a)

The well-known delta function of the Kerr metric

RelativisticDynamics.initial_conditions — Method

initialization = initial_conditions(M)

Setup the initial conditions for the MPD orbital dynamics

RelativisticDynamics.mapping_d — Method

d = mapping_d(r,a,zminus)

Mapping function d used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_f — Method

f = mapping_f(r,a,zminus)

Mapping function f used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_g — Method

g = mapping_g(r,a)

Mapping function g used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_h — Method

h = mapping_h(r,a,zminus)

Mapping function h used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.metric_contra_g11 — Method

g = metric_g11(coords,a)

The covariant tt component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_contra_g14 — Method

g = metric_g14(coords,a)

The covariant tϕ component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_contra_g22 — Method

g = metric_g22(coords,a)

The covariant rr component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_contra_g33 — Method

g = metric_g33(coords,a)

The covariant θθ component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_contra_g44 — Method

g = metric_g44(coords,a)

The covariant ϕϕ component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_g11 — Method

g = metric_g11(coords,a)

The covariant tt component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_g14 — Method

g = metric_g14(coords,a)

The covariant tϕ component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_g22 — Method

g = metric_g22(coords,a)

The covariant rr component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_g33 — Method

g = metric_g33(coords,a)

The covariant θθ component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.metric_g44 — Method

g = metric_g44(coords,a)

The covariant ϕϕ component of the Kerr metric in Boyer Lindquist coordinates

RelativisticDynamics.orbit — Method

solution,model = orbit(NF,kwargs...)

Runs RelativisticDynamics.jl with number format NF and any additional parameters in the keyword arguments kwargs.... Any unspecified parameters will use the default values as defined in src/parameters.jl.

RelativisticDynamics.permutation_tensor — Method

ϵ = permutation_tensor(metric)

Calcualte the Levi-civita tensor in an arbitrary basis.

RelativisticDynamics.riemann — Method

R = riemann(coords,a)

Riemann tensor components of the Kerr metric. First index is the contravariant, others are covariant

RelativisticDynamics.schwarzchild_covariant_riemann — Method

R = schwarzchild_covariant_riemann(coords,a)

Special case - the fully covariant components of the Riemann tensor for schwarzchild metric Used for testing

RelativisticDynamics.sigma — Method

Σ = sigma(r,θ,a)

The well-known sigma function of the Kerr metric

RelativisticDynamics.spintensor — Method

S^{μ ν} = spintensor(levi,pvector,svector,m0)

Calculate the contravariant spin tensor.

RelativisticDynamics.timestepping — Method

solution = timestepping(X,M)

The timestepping integration once all variables have been initialised