RelativisticDynamics

Documentation for RelativisticDynamics.

Welcome to the documentation for RelativisticDynamics.jl a relativistic orbital dynamics model for simulating spinning objects in curved spacetime.

RelativisticDynamics.Constants

RelativisticDynamics.Constants

RelativisticDynamics.SystemParameters

RelativisticDynamics.orbit

RelativisticDynamics.Constants — Type

C = Constants(P)

A struct to hold all variables which are constant over the course of the integration.

RelativisticDynamics.Constants — Method

Generator function for a Constants struct.

RelativisticDynamics.SystemParameters — Type

P = Parameters(kwargs...)

A struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.

RelativisticDynamics.orbit — Method

Some docstring for the run_program function