RelativisticDynamics

Documentation for RelativisticDynamics.

Welcome to the documentation for RelativisticDynamics.jl a relativistic orbital dynamics model for simulating spinning objects in curved spacetime.

RelativisticDynamics.Constants

RelativisticDynamics.Constants

RelativisticDynamics.Model

RelativisticDynamics.PrognosticVariables

RelativisticDynamics.SystemParameters

RelativisticDynamics.SphericalPhotonOrbit_initial_conditions

RelativisticDynamics.convert_to_covariant

RelativisticDynamics.delta

RelativisticDynamics.initial_conditions

RelativisticDynamics.mapping_d

RelativisticDynamics.mapping_f

RelativisticDynamics.mapping_g

RelativisticDynamics.mapping_h

RelativisticDynamics.orbit

RelativisticDynamics.sigma

RelativisticDynamics.Constants — Type

C = Constants(P)

A struct to hold all variables which are constant over the course of the integration.

RelativisticDynamics.Constants — Method

Generator function for a Constants struct.

RelativisticDynamics.Model — Type

M = BarotropicModel(::Parameters,
                    ::Constants,
                    ::GeoSpectral,
                    ::HorizontalDiffusion)

The BarotropicModel struct holds all other structs that contain precalculated constants, whether scalars or arrays that do not change throughout model integration. In contrast to ShallowWaterModel or PrimitiveEquationModel it does not contain a Boundaries struct as not needed.

RelativisticDynamics.PrognosticVariables — Type

Struct holding the so-called 'prognostic' variables

RelativisticDynamics.SystemParameters — Type

P = Parameters(kwargs...)

A struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.

RelativisticDynamics.SphericalPhotonOrbit_initial_conditions — Method

Initialize a PrognosticVariables struct for an atmosphere at rest. No winds, hence zero vorticity and divergence, but temperature, pressure and humidity are initialised

RelativisticDynamics.convert_to_covariant — Method

Convert a vector from contravariant form to convariant form using the metric

RelativisticDynamics.delta — Method

Δ = delta(r,a) The well-known delta function of the Kerr metric

RelativisticDynamics.initial_conditions — Method

Initialize a PrognosticVariables struct for an atmosphere at rest. No winds, hence zero vorticity and divergence, but temperature, pressure and humidity are initialised

RelativisticDynamics.mapping_d — Method

d = mapping_d(r,a,zminus) Mapping function d used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_f — Method

f = mapping_f(r,a,zminus) Mapping function f used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_g — Method

g = mapping_g(r,a) Mapping function g used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_h — Method

h = mapping_h(r,a,zminus) Mapping function h used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.orbit — Method

Some docstring for the run_program function

RelativisticDynamics.sigma — Method

Σ = sigma(r,θ,a) The well-known sigma function of the Kerr metric