RelativisticDynamics

Documentation for RelativisticDynamics.

Welcome to the documentation for RelativisticDynamics.jl a relativistic orbital dynamics model for simulating spinning objects in curved spacetime.

RelativisticDynamics.Constants

RelativisticDynamics.Constants

RelativisticDynamics.Model

RelativisticDynamics.PrognosticVariables

RelativisticDynamics.SystemParameters

RelativisticDynamics.ELQ

RelativisticDynamics.Kretschmann_scalar

RelativisticDynamics.MPD_initial_conditions

RelativisticDynamics.PlotTrajectory

RelativisticDynamics.bounds_checks

RelativisticDynamics.christoffel

RelativisticDynamics.contravariant_metric

RelativisticDynamics.convert_to_covariant

RelativisticDynamics.covariant_metric

RelativisticDynamics.covariant_minkowski

RelativisticDynamics.delta

RelativisticDynamics.mapping_d

RelativisticDynamics.mapping_f

RelativisticDynamics.mapping_g

RelativisticDynamics.mapping_h

RelativisticDynamics.orbit

RelativisticDynamics.riemann

RelativisticDynamics.schwarzchild_covariant_riemann

RelativisticDynamics.sigma

RelativisticDynamics.spintensor

RelativisticDynamics.Constants — Type

C = Constants(P) A struct to hold all variables which are constant over the course of the integration. These are derived from the user-defined parameters

RelativisticDynamics.Constants — Method

Generator function for a Constants struct.

RelativisticDynamics.Model — Type

RelativisticDynamics.PrognosticVariables — Type

Struct holding the so-called 'prognostic' variables

RelativisticDynamics.SystemParameters — Type

P = Parameters(kwargs...) A struct to hold all model parameters that may be changed by the user. The struct uses keywords such that default values can be changed at creation. The default values of the keywords define the default model setup.

RelativisticDynamics.ELQ — Method

Calculate the energy, angular momentum and Carter constant for the different types of system

RelativisticDynamics.Kretschmann_scalar — Method

Kretschman scalar for the Kerr metric

RelativisticDynamics.MPD_initial_conditions — Method

Setup the initial conditions for the MPD orbital dynamics

RelativisticDynamics.PlotTrajectory — Function

Plot trajectory of a body. Assumes coordinates are Boyer Lindquist. Plots in either 2D or 3D depending on specification of dimensions. Saves a low resolution PNG figure to disk in example_media/

RelativisticDynamics.bounds_checks — Method

Look before you leap

RelativisticDynamics.christoffel — Method

Christoffel symbols of the second kind

RelativisticDynamics.contravariant_metric — Method

Construct the NxN matrix of the contravariant metric. Metric components are defined via indvidual functions to allow for auto diff in unit tests

RelativisticDynamics.convert_to_covariant — Method

Convert a vector from contravariant form to convariant form using the covariant metric

RelativisticDynamics.covariant_metric — Method

Construct the NxN matrix of the covariant metric. Metric components are defined via indvidual functions to allow for auto diff in unit tests

RelativisticDynamics.covariant_minkowski — Method

Construct the NxN matrix of the covariant Minkowski metric.

RelativisticDynamics.delta — Method

Δ = delta(r,a) The well-known delta function of the Kerr metric

RelativisticDynamics.mapping_d — Method

d = mapping_d(r,a,zminus) Mapping function d used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_f — Method

f = mapping_f(r,a,zminus) Mapping function f used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_g — Method

g = mapping_g(r,a) Mapping function g used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.mapping_h — Method

h = mapping_h(r,a,zminus) Mapping function h used when converting from Keplerian orbital parameters to constants of motion

RelativisticDynamics.orbit — Method

Some docstring for the run_program function

RelativisticDynamics.riemann — Method

Riemann tensor components. First index is the contravariant, others are covariant

RelativisticDynamics.schwarzchild_covariant_riemann — Method

Special case - the fully covariant components of the Riemann tensor for schwarzchild. From https://arxiv.org/pdf/0904.4184.pdf

RelativisticDynamics.sigma — Method

Σ = sigma(r,θ,a) The well-known sigma function of the Kerr metric

RelativisticDynamics.spintensor — Method

Calculate the contravariant spin tensor